+++ cmd/mq_influx/exporter.go
@@ -232,6 +232,14 @@ func Collect(c client.Client) error {
 							series = "qmgr"
 							if key == mqmetric.QMgrMapKey {
 								tags["description"] = mqmetric.GetObjectDescription("", ibmmq.MQOT_Q_MGR)
+								hostname := mqmetric.GetQueueManagerAttribute(config.cf.QMgrName, ibmmq.MQCACF_HOST_NAME)
+								if hostname != mqmetric.DUMMY_STRING {
+									tags["hostname"] = hostname
+								}
+							} else if strings.HasPrefix(key, mqmetric.NativeHAKeyPrefix) {
+								series = "nha"
+								tags["nhainstance"] = strings.Replace(key, mqmetric.NativeHAKeyPrefix, "", -1)
+
 							} else {
 								tags["queue"] = key
 								series = "queue"
@@ -427,6 +435,10 @@ func Collect(c client.Client) error {
 								"qmgr":        qMgrName,
 								"platform":    platformString,
 								"description": mqmetric.GetObjectDescription("", ibmmq.MQOT_Q_MGR),
+							}
+							hostname := mqmetric.GetQueueManagerAttribute(config.cf.QMgrName, ibmmq.MQCACF_HOST_NAME)
+							if hostname != mqmetric.DUMMY_STRING {
+								tags["hostname"] = hostname
 							}
 
 							f := mqmetric.QueueManagerNormalise(attr, value.ValueInt64)